pyproject.toml
[project]
name = "user-api"
version = "0.1.0"
description = "API REST para criação e gerenciamento de usuários"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.135.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.13.0",
    "psycopg2-binary>=2.9.9",
    "pydantic[email]>=2.7.0",
    "dynaconf>=3.2.5",
    "bcrypt>=4.0.0",
    "dependency-injector>=4.41.0",
    "PyJWT>=2.8.0",
    "pydantic-settings>=2.13.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "httpx>=0.27.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.15.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
testpaths = ["tests"]

